Function eval_all_def added, and minor bug fixes
I added the function eval_all_def that calls multiple times the function
eval_def_once. This function evaluates all def directive and produces an
error if a def directive can't be evaluated.
The bug fixes are in the function eval_def_once and in rpneval.
After a def directive is evaluated, the code of the instruction is
changed to 0. Before, it was changed to 'no_mnemonic' which has a value
of 18. 'no_mnemonic' corresponds to the family of instruction and not to
the instruction's code.
In the function rpneval, I added a buffer in order to copy the
expression before passing it to strtok. So if an rpn expression has to
be evaluated a second time, it won't modify the input expression.